Basic Interaction System (Incomplete)
Implemented a basic interaction system in the mushroom garden scene only. The system allows for the swapping of Interaction ScriptableObjects for each interaction trigger (interactable), enabling the playing of different Interactions at different points in the game (which interaction to play will be managed by a Quest Manager).

Issues not yet fixed:
- Player doesn't pause when interaction is playing
- Haven't tested system extensively with actual quests/more complex interactions